requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
DateTime==5.4
debtcollector==3.0.0
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
dnspython==2.2.1
email-validator==1.3.0
eventlet==0.35.2
executing==2.0.1
fasteners==0.19
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.48.1
greenlet==3.0.3
gunicorn==20.0.4
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
ipykernel==6.29.0
ipython==8.20.0
iso8601==2.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
load-dotenv==0.1.0
Mako==1.3.2
MarkupSafe==2.1.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
msgpack==1.0.7
nest-asyncio==1.5.9
netaddr==1.2.1
netifaces==0.11.0
numpy==1.26.4
openai==1.12.0
os-win==5.9.0
oslo.concurrency==6.0.0
oslo.config==9.4.0
oslo.context==5.4.0
oslo.i18n==6.3.0
oslo.log==5.5.0
oslo.serialization==5.4.0
oslo.utils==7.1.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pbr==6.0.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pyodbc==5.1.0
pyparsing==3.1.1
pypyodbc==1.3.6
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.28.1
rfc3986==2.0.0
seaborn==0.13.2
setuptools==69.1.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
stack-data==0.6.3
stevedore==5.2.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.12
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.0.1
zope.interface==6.2
